[gin-tuto] 메인코드로 정리

+++ gin-tuto/model/response.go
@@ -3,7 +3,7 @@ package model
 type StResponse struct {
 	Type     string `json:"type"`
 	Title    string `json:"title"`
-	Status   int64  `json:"status"`
+	Status   int    `json:"status"`
 	Detail   string `json:"detail"`
 	Instance string `json:"instance"`
 }
